+++ common/gpu.h
@@ -8,8 +8,8 @@ void gpu_init(void);
 void gpu_reset(void);
 void gpu_done(void);
 void gpu_set_map_func(SV_MapRGBFunc func);
-int  gpu_get_color_scheme(void);
-void gpu_set_color_scheme(int colorScheme);
+int8 gpu_get_color_scheme(void);
+void gpu_set_color_scheme(int8 colorScheme);
 void gpu_render_scanline(uint32 scanline, uint16 *backbuffer, uint8 innerx, uint8 size);
 void gpu_set_ghosting(int frameCount);
